// webapp/src/app/home/types/cart.types.ts
import { Product } from "./products.types";

export interface CartItem {
    product: Product;
    quantity: number;
    amount: number;
}

export interface Cart {
    products: CartItem[];
    totalAmount: number;
    totalProducts: number;
}

export interface DeliveryAddress {
    userName: string;
    address: string;
    city: string;
    state: string;
    pin: string;
}